+++ src/headers/shared.h
@@ -185,7 +185,10 @@ extern const char *__local_name;
 
 #define os_strdup(x,y) ((y = strdup(x)))?(void)1:merror_exit(MEM_ERROR, errno, strerror(errno))
 
+/* Windows doesn't support strndup function */
+#ifndef WIN32
 #define os_strndup(x,y,z) ((z = strndup(x,y)))?(void)1:merror_exit(MEM_ERROR, errno, strerror(errno))
+#endif
 
 #define os_malloc(x,y) ((y = (__typeof__(y)) malloc(x)))?(void)1:merror_exit(MEM_ERROR, errno, strerror(errno))